Change project structure on backend Change names in all files on backend
