+++ Backend/src/middleware/auth-middleware.ts
