+++ Backend/src/middleware/image-middleware.ts
