+++ Backend/src/routes/mail-route.ts
@@ -1,8 +1,8 @@
 import { Router } from 'express';
-import MailController from '../controllers/mailController';
+import MailController from '../controllers/mail-controller';
 import RouterWithPath from '../types/router';
-import validationMiddleware from '../middleware/validationMiddleware';
-import { mailSchema } from '../validations/mail';
+import validationMiddleware from '../middleware/validation-middleware';
+import { mailSchema } from '../controllers/validations/mail-validation';
 
 class MailRoutes implements RouterWithPath {
   public path = '/message';
