+++ Backend/src/server.ts
@@ -2,14 +2,14 @@ import 'dotenv/config';
 import validateEnv from './utils/validateEnv';
 import App from './app';
 
-import PostRoutes from './routes/postRoutes';
-import ProjectRoutes from './routes/projectRoutes';
+import PostRoutes from './routes/posts-route';
+import ProjectRoutes from './routes/projects-route';
 import AuthenticationRoutes from './routes/authentication-route';
-import CVRoutes from './routes/cvRoutes';
-import MailRoutes from './routes/mailRoutes';
-import SkillsRoutes from './routes/skillRoutes';
+import CVRoutes from './routes/cv-route';
+import MailRoutes from './routes/mail-route';
+import SkillsRoutes from './routes/skills-route';
 import createStorage from './utils/createStorage';
-import ImageRoutes from './routes/imageRoutes';
+import ImageRoutes from './routes/image-route';
 
 validateEnv();
 createStorage(['./uploads', './uploads/images']);
